fix test case
