+++ sofa-ark-parent/support/ark-springboot-integration/ark-springboot-starter/src/test/java/com/alipay/sofa/ark/test/springboot/RegisterMockEmbedTomcatService.java
@@ -1,0 +1,42 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.alipay.sofa.ark.test.springboot;
+
+import com.alipay.sofa.ark.container.registry.ContainerServiceProvider;
+import com.alipay.sofa.ark.spi.service.ArkInject;
+import com.alipay.sofa.ark.spi.service.registry.RegistryService;
+import com.alipay.sofa.ark.spi.web.EmbedTomcatService;
+import com.alipay.sofa.ark.web.embed.tomcat.EmbedTomcatServiceImpl;
+import org.springframework.beans.factory.InitializingBean;
+import org.springframework.beans.factory.config.BeanPostProcessor;
+import org.springframework.stereotype.Component;
+
+/**
+ * @author qilong.zql
+ * @since 0.6.0
+ */
+@Component
+public class RegisterMockEmbedTomcatService implements BeanPostProcessor, InitializingBean {
+    @ArkInject
+    private RegistryService registryService;
+
+    @Override
+    public void afterPropertiesSet() throws Exception {
+        registryService.publishService(EmbedTomcatService.class, new EmbedTomcatServiceImpl(),
+            new ContainerServiceProvider());
+    }
+}
