+++ sofa-ark-parent/support/ark-springboot-integration/ark-springboot-starter/src/test/resources/META-INF/sofa-ark-test/sofa-ark-test.xml
@@ -3,8 +3,8 @@
        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
        xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
 
-    <bean id="sampleService" class="com.alipay.sofa.ark.springboot.impl.SampleServiceImpl"/>
+    <bean id="sampleService" class="com.alipay.sofa.ark.test.springboot.impl.SampleServiceImpl"/>
 
-    <bean id="testBizEventHandler" class="com.alipay.sofa.ark.springboot.impl.TestBizEventHandler"/>
+    <bean id="testBizEventHandler" class="com.alipay.sofa.ark.test.springboot.impl.TestBizEventHandler"/>
 
 </beans>
